Refactored codebase for modularity, maintainability, and frontend integration preparation
[x] - Moved image description logic to a dedicated visionService
[x] - Created cacheService for handling cache directory and file operations
[x] - Ensured proper logging in cacheService for debugging and monitoring
[x] - Updated newsletterService to utilize visionService and cacheService
[x] - Added human-readable logging to fetchAndProcessMessages for better debugging
[x] - Simplified and streamlined the process of fetching, processing, and caching messages
[x] - Enhanced modular structure in preparation for frontend integration
[x] - Decoupled logic from routes, focusing on controller-like structure for handling requests
[x] - Implemented proper environment variable management with dotenv
[x] - Added cron jobs for automated weekly newsletter generation and sending
[x] - Integrated OpenAI and Google Vision API through gateway services for extensibility
[x] - Consolidated app initialization logic in index.js for better organization
[x] - Verified WhatsApp client session persistence to avoid frequent reconnections
[x] - Implemented error handling and logging across services for robust operation
[x] - Added tasks for upcoming features: dynamic frontend generation with React and MUI, voice message processing, and comprehensive context handling for better summaries

Upcoming: Dynamic frontend generation using React for a modern, vibrant newsletter presentation
