--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "whatsapp_newsletter_bot",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@google-cloud/vision": "^4.3.0",
-    "axios": "^1.7.2",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "node-cron": "^3.0.3",
-    "openai": "^4.51.0",
-    "puppeteer": "^22.11.2",
-    "puppeteer-core": "^22.11.2",
-    "qrcode-terminal": "^0.12.0",
-    "whatsapp-web.js": "^1.24.0"
-  },
-  "type": "module"
-}
