--- services/helpers/ensureCacheDirExists.js
@@ -1,15 +1,0 @@
-import fs from 'fs';
-import path from 'path';
-import { fileURLToPath } from 'url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-const ensureCacheDirExists = () => {
-    const cacheDir = path.join(__dirname, '../cache');
-    if (!fs.existsSync(cacheDir)) {
-        fs.mkdirSync(cacheDir, { recursive: true });
-    }
-};
-
-export default ensureCacheDirExists;
